a function given as a parameter on each element of an array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,10 +11,10 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (array == NULL || size < 1 || action == NULL)
-	return;
+		return;
 
 	for (i = 0; i < size; i++)
 	{
